pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "segment_tools"
version = "1.0.0"
authors = [  { name="roiro" },]
description = "Segmentation tools"
license = {text = "MIT License"}
requires-python = "==3.10.*"
dependencies = [
    "matplotlib>=3.2.2",
    "opencv-python>=4.6.0",
    "Pillow>=7.1.2",
    "PyYAML>=5.3.1",
    "requests>=2.23.0",
    "scipy>=1.4.1",
    "torch>=1.7.0",
    "torchvision>=0.8.1",
    "tqdm>=4.64.0",
    "pandas>=1.1.4",
    "seaborn>=0.11.0",
    "gradio==3.35.2",
    "ultralytics == 8.0.120",
    "clip @ git+https://github.com/openai/CLIP.git",
    "GroundingDINO @ git+https://github.com/qwertyroiro/GroundingDINO_GSA",
    "segment_anything @ git+https://github.com/qwertyroiro/segment_anything_GSA",
    "transformers",
    "addict",
    "diffusers",
    "huggingface_hub",
    "numpy",
    "onnxruntime",
    "pycocotools",
    "requests",
    "setuptools",
    "supervision",
    "termcolor",
    "timm",
    "transformers",
    "yapf",
    "nltk",
    "fairscale",
    "litellm"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.hatch.metadata]
allow-direct-references = true